DPSTAT-834: Integration test (#339)
Add integration tests for Pseudonymize, Validator and Depseudonymize

This PR adds integration tests for Pseudonymize, Validator and Depseudonymize. These tests are executed against the `dapla-pseudo-service` in a staging. As of now the tests only covers the basic usage of toolbelt-pseudo, this should be improved as we further develop toolbelt-pseudo.

**Local**: When tests run locally they use the user's Google credentials for authorization. This requires that the user is part of the service admin group in the staging. For example, users can be added to the service configuration as demonstrated in this PR: https://github.com/statisticsnorway/platform-dev/pull/3957
**GH Actions**: The tests are setup in it's own workflow `pseudo-service-integration-test`, it uses the SA `gha-pseudo-integration-test-sa` for authorization with `dapla-pseudo-service` in staging, and then runs the tests.

When adding additional integration tests they should to be annotated with `@integration_test` and passed the `setup` fixture. This ensures that tests are configured to use `dapla-pseudo-service` in a staging.

**Setup of integration test SA**:
- https://github.com/statisticsnorway/dapla-pseudo-iac/pull/19
- https://github.com/statisticsnorway/dapla-pseudo-iac/pull/20
- https://github.com/statisticsnorway/dapla-pseudo-iac/pull/21